First Version of Infinite Snake
This first Version of infinite Snake is rather unpolished and unoptimised even though it runs very smoothly with practically no input or rendering lag.

For the best experience you should play on a 1080p Monitor. That the surface behaves as if it was one of a donut means, that you can move out any side and come back in in the opposite.